fix: add PasswordInput component
